refactor(firmware): move sketch headers/sources into include/src
* Rename the sketch's own headers from `.h` to `.hpp` and move them into
  `include/`: `pin_config`, `dex`, `species`, `pet`, `audio`, `rtcbat`,
  `sdmon`, `i18n`.
* Move the matching `.cpp` sources into `src/`, relying on Arduino's `src`
  sketch convention (compiled recursively, not shown as IDE tabs).
* Update every local `#include` accordingly: the `.ino` now references
  `include/*.hpp`, and each `src/*.cpp` references `../include/*.hpp`
  (Arduino does not add either folder to the include search path on its
  own, so paths must be explicit).
* Update `tools/gen_dex.py` and `tools/sprites.py` to write their generated
  headers to `include/dex.hpp` and `include/species.hpp`.
* Update README's layout section and file references to match.

Verified with a full `arduino-cli compile` against the FQBN documented in
the README (esp32:esp32:esp32s3, PSRAM=opi, 16M flash): builds clean,
596896 bytes program / 33656 bytes globals, no new warnings.
